requirements.txt
leglight
obs-websocket-py
streamdeck
